# docs/openapi.yml
openapi: 3.1.0
info:
  title: fast-auth
  description: A simple, extensible authentication library for Axum with JWT and refresh tokens
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
...
  /auth/password/forgot:
    post:
      tags:
      - auth
      summary: Request a password reset email.
      description: |-
        Creates a reset token and sends an email with the reset link.
        Always returns success to prevent email enumeration attacks.
      operationId: password_forgot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordForgotRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordForgotResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/password/reset:
    post:
      tags:
      - auth
      summary: Reset password using the token from the email.
      operationId: password_reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh the current session using the refresh-token cookie.
      description: |-
        Rotates the refresh token, issues a new access token, and returns the
        refreshed user payload.
      operationId: refresh
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
...
components:
  schemas:
...
    AuthErrorResponse:
      type: object
      description: Standard error response body returned by auth endpoints.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
    AuthResponse:
      type: object
      description: Auth response body. Tokens are set as httpOnly cookies.
      required:
      - user
      - organization
      - auth_role
      properties:
        auth_role:
          type: string
        organization:
          $ref: '#/components/schemas/OrganizationResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
...
    OrganizationResponse:
      type: object
      description: Active organization returned in auth responses.
      required:
      - id
      - name
      - role
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/OrganizationRole'
    OrganizationRole:
      type: string
      description: |-
        Organization membership role exposed by auth.

        This role is embedded in [`CurrentUser`], organization membership responses,
        and invitations so backends and HTTP handlers can share one stable
        authorization vocabulary.
      enum:
      - owner
      - admin
      - member
...
    PasswordForgotRequest:
      type: object
      description: Request body for forgot password.
      required:
      - email
      properties:
        email:
          type: string
          description: User's email address.
    PasswordForgotResponse:
      type: object
      description: Response for forgot password.
      required:
      - message
      properties:
        message:
          type: string
          description: Success message.
    PasswordResetRequest:
      type: object
      description: Request body for password reset.
      required:
      - token
      - password
      properties:
        password:
          type: string
          description: New password.
        token:
          type: string
          description: Reset token from email link.
    PasswordResetResponse:
      type: object
      description: Response for password reset.
      required:
      - message
      properties:
        message:
          type: string
          description: Success message.
...
    UserResponse:
      type: object
      description: User data returned in auth responses.
      required:
      - id
      - email
      properties:
        email:
          type: string
        email_confirmed_at:
          type:
          - string
          - 'null'
        id:
          type: string
  securitySchemes:
    bearerApiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
...